refactor(test): P5-1 follow-up — extract Wave 2 cheap-wiring bundle
Continue test_widget_bridge.cpp split. Extracts the 731-line Wave 2
cheap-wiring bundle into a sibling TU.

Two compat.json closure clusters unified by being one-line runtime
wiring tests:
  1. canvas2d cheap wiring (DIVERGE → PASS) — bridge ↔ Canvas API
     contract for previously-unwired CanvasRenderingContext2D
     entries.
  2. css bundle cheap value-coverage wiring — JS shim → CSS
     translator → View slot for unwired CSS property/value
     combinations.

* test_widget_bridge.cpp: 6,214 → 5,483 lines (-731)
* New target pulp-test-widget-bridge-wave2-cheap: 780 lines,
  18 TEST_CASEs, 93 assertions.

This is commit 2 of 2 in a bundled-cuts PR (yoga + A4 OOS shipped
in commit 1; single version bump comes next).
